reverted, but with fixes.
In the previous commit, I tried to change the calling conventions for the functions, and it caused a major cascade of errors. 
And, there were a few rounding bugs that slipped through. Replaced with the working, albeit bad calling conventions, and fixed the rounding bugs. 
Began work on expanding the the signs api to be generic enough to allow for dynamic sign creations in other mods by calling the api.